feat: add support for WAR_PATH env var
Add support for a new `WAR_PATH` environment variable that allows users to
specify where the .war file too run is stored. This is useful when the .war
file is not stored at the root of the application.

Also downloads JVM common tools to cache, and use the cached version if
it's available.

Also updates the default values for:
- `JAVA_VERSION`: from `1.8` to `17`
- `WEBAPP_RUNNER_VERSION`: from `8.5.11.3` to `9.0.52.1`

I'm also suggesting to rename `WEBAPP_RUNNER_VERSION` to
`JAVA_WEBAPP_RUNNER_VERSION`. The code still supports
`WEBAPP_RUNNER_VERSION`.

The README file has been updated consequently.